fix: Kokoro speed tensor type + Supertonic debug diagnostics
- Fix Kokoro speed tensor: was i64 (integer), model expects f32 (float)
- Add debug eprintln to Supertonic: model input names, preprocessed text,
  token IDs, duration, latent dims, vocoder output shape
- Use zeros instead of random noise for deterministic debugging
- Fix text_emb_shape to use usize (was i64 from try_extract_tensor)

+++ src/engine/kokoro.rs
@@ -94,7 +94,7 @@ impl KokoroEngine {
         let style_value = Value::from_array(([1usize, 256usize], style_slice.to_vec()))
             .with_context(|| "Failed to create style tensor")?;
 
-        let speed_value = Value::from_array(([1usize], vec![self.speed as i64]))
+        let speed_value = Value::from_array(([1usize], vec![self.speed]))
             .with_context(|| "Failed to create speed tensor")?;
 
         let outputs = self
